+++ lib/loops_sdk/audience_segments.rb
@@ -10,6 +10,11 @@ module LoopsSdk
       def get(audience_segment_id:)
         make_request(method: :get, path: "v1/audience-segments/#{audience_segment_id}")
       end
+
+      def create(name:, filter:, description: nil)
+        body = { name: name, filter: filter, description: description }.compact
+        make_request(method: :post, path: "v1/audience-segments", body: body)
+      end
     end
   end
 end