+++ lib/loops_sdk/components.rb
@@ -10,6 +10,15 @@ module LoopsSdk
       def get(component_id:)
         make_request(method: :get, path: "v1/components/#{component_id}")
       end
+
+      def create(name:, lmx:)
+        make_request(method: :post, path: "v1/components", body: { name: name, lmx: lmx })
+      end
+
+      def update(component_id:, name: nil, lmx: nil)
+        body = { name: name, lmx: lmx }.compact
+        make_request(method: :post, path: "v1/components/#{component_id}", body: body)
+      end
     end
   end
 end