+++ lib/loops_sdk/themes.rb
@@ -10,6 +10,16 @@ module LoopsSdk
       def get(theme_id:)
         make_request(method: :get, path: "v1/themes/#{theme_id}")
       end
+
+      def create(name:, styles: nil)
+        body = { name: name, styles: styles }.compact
+        make_request(method: :post, path: "v1/themes", body: body)
+      end
+
+      def update(theme_id:, name: nil, styles: nil)
+        body = { name: name, styles: styles }.compact
+        make_request(method: :post, path: "v1/themes/#{theme_id}", body: body)
+      end
     end
   end
 end